[NO-TICKET] Add script for finding React versions used by dependents (#1849)
* Add script for finding react versions

* Turn them into one command that can execute different dependent scans

* Move the readme and delete old files

* This doesn't need to be a separate file unless the fn is reused

* Move error handling to the root

* Simplify file names

* Remove unused import

* Remove commented-out code

* Explain the token env var better
