+++ package.json
@@ -44,7 +44,7 @@
     "gh-pages": "yarn gh-pages:healthcare && yarn gh-pages:medicare",
     "gh-pages:healthcare": "yarn build-docs:healthcare && gh-pages -d './packages/ds-healthcare-gov/docs/dist' --dest 'healthcare'",
     "gh-pages:medicare": "yarn build-docs:medicare && gh-pages -d './packages/ds-medicare-gov/docs/dist' --dest 'medicare'",
-    "scan-versions": "ts-node scripts/findDesignSystemVersions/index.ts",
+    "scan-dependents": "ts-node scripts/scanDependents/index.ts",
     "token:dist": "yarn --cwd ./packages/design-system-tokens build themes scss && yarn --cwd ./packages/design-system-tokens dist"
   },
   "devDependencies": {
